UPDATE: the js is now integrated inside the GlobalStatsTag (self-contained)
